Factor out RecordsDatetimeFixture
